# Enhance performance of arrayProxy of primitives.