Summary:
- There was 30% gap between VectorBasic and SimpleProxyWithResize
in the case of no nulls. This diff reduces the gap to 18%.

- The change also makes the "with nulls" case slightly faster.

- The change simplify the construction of the PrimitiveWriterProxy.
- The diff also remove FOLLY_ALWAYS_INLINE from ArrayProxy since it
 did not  effect performance.
```
no nulls:

before:

VectorBasic                                                160.02ns    6.25M
VectorResizeOptimized                                       76.78ns   13.02M
SimpleProxyWithResize                                      207.59ns    4.82M

after:

VectorBasic                                                160.02ns    6.25M
VectorResizeOptimized                                       76.78ns   13.02M
SimpleProxyWithResize                                      188.90ns    5.29M
```

```
with nulls
after:

VectorBasic                                                405.50ns    2.47M
VectorResizeOptimized                                      313.68ns    3.19M
SimpleProxyWithResize                                      392.09ns    2.55M
```

Differential Revision: D33715941